start a video-only page; fix some styles
