+++ webroot/index.html
@@ -33,6 +33,7 @@
     <link href="//unpkg.com/video.js@7.9.2/dist/video-js.css" rel="stylesheet">
     <link href="https://unpkg.com/@videojs/themes@1/dist/fantasy/index.css" rel="stylesheet" />
     <script src="//unpkg.com/video.js@7.9.2/dist/video.js"></script>
+    <!-- markdown renderer -->
     <script src="//unpkg.com/showdown/dist/showdown.min.js"></script>
 
     <link href="./styles/app.css" rel="stylesheet" />
@@ -48,7 +49,7 @@
 
   <script type="module">
     import { render, html } from "https://unpkg.com/htm/preact/index.mjs?module";
-    import App from './js/app2.js';
+    import App from './js/app.js';
 
     (function () {
       render(html`<${App} />`, document.getElementById("app"));
